[Refactor] Modify test code
